refactor(mvp): resolve the perspective from a [perspective] root segment
Moves the website routes under `app/[perspective]`, adds `proxy.ts` with
`definePerspectiveProxy`, and a shared `app/sanity.live.ts` that hands
`defineLive` the root param getter. Pages render a single `'use cache'`
leaf inside Suspense and call `sanityFetch({query})` with no perspective
or stega props. The no-resolver page keeps its own strict `defineLive`
and passes `perspective: 'drafts'` to show the other contract.
`getDynamicFetchOptions` is deleted, the debug status widget still reads
the preview cookies for display.
