+++ apps/mvp/app/[perspective]/DebugStatus.tsx
