+++ apps/mvp/app/[perspective]/no-resolve-perspective/page.tsx
@@ -1,68 +1,47 @@
-import {defineLive, type StrictDefinedFetchType} from 'next-sanity/live'
-import {draftMode} from 'next/headers'
+import {defineLive} from 'next-sanity/live'
 import Link from 'next/link'
 import {Suspense} from 'react'
 
-import PostsLayout, {postsQuery} from '@/app/(website)/PostsLayout'
+import PostsLayout, {postsQuery} from '@/app/[perspective]/PostsLayout'
 import {client} from '@/app/sanity.client'
 
 import {ContentSourceMapDebug} from '../ContentSourceMapDebug'
-import {getDynamicFetchOptions, type DynamicFetchOptions} from '../resolvePreviewCookies'
+import {onError} from './client-functions'
 
 const token = process.env.SANITY_API_READ_TOKEN!
-const {sanityFetch} = defineLive({
+const {sanityFetch, SanityLive} = defineLive({
   client,
   serverToken: token,
   browserToken: process.env.NEXT_PUBLIC_SANITY_API_BROWSER_TOKEN || token,
   strict: true,
 })
 
-// The app's one shared 'use cache' boundary. `sanityFetch` calls
-// `cacheTag`/`cacheLife` internally but doesn't create the boundary —
-// this wrapper provides it once, so callers don't add their own.
-const cachedSanity: StrictDefinedFetchType = async (options) => {
+async function CachedIndexPage() {
   'use cache'
-  return sanityFetch(options)
-}
-
-async function CachedIndexPage({perspective, variant, stega}: DynamicFetchOptions) {
-  const {data, sourceMap, tags} = await cachedSanity({
+  // Without a perspective resolver the caller names the draft mode perspective.
+  // Outside draft mode strict mode forces 'published' anyway.
+  const {data, sourceMap, tags} = await sanityFetch({
     query: postsQuery.query,
-    perspective,
-    variant,
-    stega,
+    perspective: 'drafts',
   })
 
   return (
     <>
       <ContentSourceMapDebug sourceMap={sourceMap} />
-      <p>{JSON.stringify({perspective, variant, tags: tags.toSorted()})}</p>
+      <p>{JSON.stringify({tags: tags.toSorted()})}</p>
       <PostsLayout data={data} draftMode={false} />
     </>
   )
 }
 
-async function DynamicIndexPage() {
-  const {perspective, variant, stega} = await getDynamicFetchOptions()
-
-  return <CachedIndexPage perspective={perspective} variant={variant} stega={stega} />
-}
-
-export default async function IndexPage() {
-  const {isEnabled: isDraftMode} = await draftMode()
+export default function IndexPage() {
   return (
     <>
-      <div
-        className="relative bg-gray-50 px-4 pt-16 pb-20 sm:px-6 lg:px-8 lg:pt-24 lg:pb-28"
-      >
+      <div className="relative bg-gray-50 px-4 pt-16 pb-20 sm:px-6 lg:px-8 lg:pt-24 lg:pb-28">
         <div className="relative mx-auto max-w-7xl">
-          {isDraftMode ? (
-            <Suspense>
-              <DynamicIndexPage />
-            </Suspense>
-          ) : (
-            <CachedIndexPage perspective="published" stega={false} />
-          )}
+          <Suspense>
+            <CachedIndexPage />
+          </Suspense>
         </div>
       </div>
       <div className="flex gap-2 text-center">
@@ -73,13 +52,9 @@ export default async function IndexPage() {
         >
           Resolve perspective
         </Link>
-        <Link
-          prefetch={false}
-          href="/no-resolve-perspective"
-          className="mx-2 my-4 inline-block rounded-full border border-gray-200 px-4 py-1 text-sm font-semibold text-gray-600 hover:border-transparent hover:bg-gray-600 hover:text-white focus:ring-2 focus:ring-gray-600 focus:ring-offset-2 focus:outline-hidden"
-        >
+        <span className="mx-2 my-4 inline-block rounded-full border border-transparent bg-gray-600 px-4 py-1 text-sm font-semibold text-white">
           No resolve perspective
-        </Link>
+        </span>
         <Link
           prefetch={false}
           href="/only-production"
@@ -87,9 +62,13 @@ export default async function IndexPage() {
         >
           Only production
         </Link>
-        <span className="mx-2 my-4 inline-block rounded-full border border-transparent bg-gray-600 px-4 py-1 text-sm font-semibold text-white">
+        <Link
+          prefetch={false}
+          href="/only-visual-editing"
+          className="mx-2 my-4 inline-block rounded-full border border-gray-200 px-4 py-1 text-sm font-semibold text-gray-600 hover:border-transparent hover:bg-gray-600 hover:text-white focus:ring-2 focus:ring-gray-600 focus:ring-offset-2 focus:outline-hidden"
+        >
           Only Visual Editing
-        </span>
+        </Link>
         <Link
           prefetch={false}
           href="/studio"
@@ -98,6 +77,7 @@ export default async function IndexPage() {
           Open Studio
         </Link>
       </div>
+      <SanityLive onError={onError} />
     </>
   )
 }
