+++ apps/mvp/app/[perspective]/only-production/page.tsx
@@ -1,7 +1,7 @@
 import {defineLive, type DefinedFetchType} from 'next-sanity/live'
 import Link from 'next/link'
 
-import PostsLayout, {postsQuery} from '@/app/(website)/PostsLayout'
+import PostsLayout, {postsQuery} from '@/app/[perspective]/PostsLayout'
 import {client} from '@/app/sanity.client'
 
 import {ContentSourceMapDebug} from '../ContentSourceMapDebug'
@@ -24,9 +24,7 @@ export default async function IndexPage() {
   return (
     <>
       <ContentSourceMapDebug sourceMap={sourceMap} />
-      <div
-        className="relative bg-gray-50 px-4 pt-16 pb-20 sm:px-6 lg:px-8 lg:pt-24 lg:pb-28"
-      >
+      <div className="relative bg-gray-50 px-4 pt-16 pb-20 sm:px-6 lg:px-8 lg:pt-24 lg:pb-28">
         <div className="relative mx-auto max-w-7xl">
           <PostsLayout data={data} draftMode={false} />
         </div>
