+++ apps/mvp/app/[perspective]/only-visual-editing/page.tsx
@@ -1,46 +1,32 @@
-import {defineLive, type StrictDefinedFetchType} from 'next-sanity/live'
-import {draftMode} from 'next/headers'
 import Link from 'next/link'
+import {Suspense} from 'react'
 
-import PostsLayout, {postsQuery} from '@/app/(website)/PostsLayout'
-import {client} from '@/app/sanity.client'
+import PostsLayout, {postsQuery} from '@/app/[perspective]/PostsLayout'
+import {sanityFetch} from '@/app/sanity.live'
 
 import {ContentSourceMapDebug} from '../ContentSourceMapDebug'
-import {onError} from './client-functions'
 
-const token = process.env.SANITY_API_READ_TOKEN!
-const {sanityFetch, SanityLive} = defineLive({
-  client,
-  serverToken: token,
-  browserToken: process.env.NEXT_PUBLIC_SANITY_API_BROWSER_TOKEN || token,
-  strict: true,
-})
-
-// The app's one shared 'use cache' boundary. `sanityFetch` calls
-// `cacheTag`/`cacheLife` internally but doesn't create the boundary —
-// this wrapper provides it once, so callers don't add their own.
-const cachedSanity: StrictDefinedFetchType = async (options) => {
+async function CachedIndexPage() {
   'use cache'
-  return sanityFetch(options)
-}
-
-export default async function IndexPage() {
-  const isDraftMode = (await draftMode()).isEnabled
-  const perspective = isDraftMode ? 'drafts' : 'published'
-  const {data, sourceMap} = await cachedSanity({
-    query: postsQuery.query,
-    perspective,
-    stega: perspective !== 'published',
-  })
+  const {data, sourceMap, tags} = await sanityFetch({query: postsQuery.query})
 
   return (
     <>
       <ContentSourceMapDebug sourceMap={sourceMap} />
-      <div
-        className="relative bg-gray-50 px-4 pt-16 pb-20 sm:px-6 lg:px-8 lg:pt-24 lg:pb-28"
-      >
+      <p>{JSON.stringify({tags: tags.toSorted()})}</p>
+      <PostsLayout data={data} draftMode={false} />
+    </>
+  )
+}
+
+export default function IndexPage() {
+  return (
+    <>
+      <div className="relative bg-gray-50 px-4 pt-16 pb-20 sm:px-6 lg:px-8 lg:pt-24 lg:pb-28">
         <div className="relative mx-auto max-w-7xl">
-          <PostsLayout data={data} draftMode={isDraftMode} />
+          <Suspense>
+            <CachedIndexPage />
+          </Suspense>
         </div>
       </div>
       <div className="flex gap-2 text-center">
@@ -51,9 +37,13 @@ export default async function IndexPage() {
         >
           Resolve perspective
         </Link>
-        <span className="mx-2 my-4 inline-block rounded-full border border-transparent bg-gray-600 px-4 py-1 text-sm font-semibold text-white">
+        <Link
+          prefetch={false}
+          href="/no-resolve-perspective"
+          className="mx-2 my-4 inline-block rounded-full border border-gray-200 px-4 py-1 text-sm font-semibold text-gray-600 hover:border-transparent hover:bg-gray-600 hover:text-white focus:ring-2 focus:ring-gray-600 focus:ring-offset-2 focus:outline-hidden"
+        >
           No resolve perspective
-        </span>
+        </Link>
         <Link
           prefetch={false}
           href="/only-production"
@@ -61,13 +51,9 @@ export default async function IndexPage() {
         >
           Only production
         </Link>
-        <Link
-          prefetch={false}
-          href="/only-visual-editing"
-          className="mx-2 my-4 inline-block rounded-full border border-gray-200 px-4 py-1 text-sm font-semibold text-gray-600 hover:border-transparent hover:bg-gray-600 hover:text-white focus:ring-2 focus:ring-gray-600 focus:ring-offset-2 focus:outline-hidden"
-        >
+        <span className="mx-2 my-4 inline-block rounded-full border border-transparent bg-gray-600 px-4 py-1 text-sm font-semibold text-white">
           Only Visual Editing
-        </Link>
+        </span>
         <Link
           prefetch={false}
           href="/studio"
@@ -76,7 +62,6 @@ export default async function IndexPage() {
           Open Studio
         </Link>
       </div>
-      <SanityLive includeDrafts={isDraftMode} onError={onError} />
     </>
   )
 }
