+++ apps/mvp/app/sanity.live.ts
@@ -1,0 +1,15 @@
+import {defineLive} from 'next-sanity/live'
+import {perspective} from 'next/root-params'
+
+import {client} from '@/app/sanity.client'
+
+const token = process.env.SANITY_API_READ_TOKEN!
+
+export const {sanityFetch, SanityLive} = defineLive({
+  client,
+  serverToken: token,
+  // TODO: setup experimental_taintUniqueValue here
+  browserToken: process.env.NEXT_PUBLIC_SANITY_API_BROWSER_TOKEN || token,
+  strict: true,
+  perspective,
+})
