+++ apps/mvp/proxy.ts
@@ -1,0 +1,10 @@
+import {definePerspectiveProxy} from 'next-sanity/live/proxy'
+
+export const proxy = definePerspectiveProxy()
+
+// Next.js requires the matcher to be a literal in this file.
+export const config = {
+  matcher: [
+    '/((?!_next|_vercel|api|studio|favicon|\\.well-known|robots\\.|sitemap\\.|[^/]*\\.).*)?',
+  ],
+}
